\name{Ant}
\alias{Ant}
\docType{data}
\title{
Ant data
}
\description{
  Species \dQuote{incidence frequency} for ant data provided by Longino et al. (2002). The ant data which used malaise traps method is collected in a tropical rain forest of Costa Rica. In the ant data set presented 455 species occurrences are distributed among 62 samples with 103 species observed. 
}
\usage{data(Ant)}
\format{
  The first entry is the total number of sampling units, and followed by the incidences frequency. The format is: \cr
 c(62, 1, 1, 1, 1, 1, ..., 19, 19, 20, 29)
}
\source{
Longino, J.T., Coddington, J. & Colwell, R.K. (2002) The ant fauna of a tropical rain forest: estimating species richness three different ways. \emph{Ecology}, \bold{83}, 689-702.
}
\examples{
data(Ant)
}
\keyword{datasets}
