\name{ChaoSpecies-package}
\alias{ChaoSpecies-package}
\docType{package}
\title{
Statistical package "\pkg{ChaoSpecies}"

}
\description{
The purpose of this R package \pkg{ChaoSpecies} is mainly to provide some species estimators for both individual-based (abundance) data and sample-based (incidence) data.
}
\details{
\tabular{ll}{
Package: \tab ChaoSpecies\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-14\cr
License: \tab GPL-2\cr
URL: \tab \url{http://chao.stat.nthu.edu.tw/blog/}
}
functions : \kbd{ChaoSpecies}, \kbd{SpeciesAbundance}, \kbd{SpeciesIncidence}
}
\author{
K.S. Tseng ; Y.H. Lee ; Anne Chao \cr
Maintainer : Anne Chao \email{chao@stat.nthu.edu.tw}
}
\keyword{SPECIES}