\name{Seedlings_Freq}
\alias{Seedlings_Freq}
\docType{data}
\title{
Frequencies Data for Multiple Samples.
}
\description{
Seedlings frequencies data is from Colwell and Coddington (1994) for illustration which contains 121 standardized soil samples collected from one-hundred 10 m x 10 m frids in a Coasta Rican forest.
}
\usage{data(Seedlings_Freq)}
\format{
The data is arranged in a matrix with 34 rows and 121 columns. Each data entry in the data matrix denotes the sample abundance or frequency. \cr
A total of 34 species of seedlings germinated from these 121 samples.
}
\details{
The format is similar to the matrix for presence/absence data except that each data entry in the data matrix denotes the sample abundance or frequency instead of incidence.
}
\source{
Colwell, R. A., and Coddington, J. A. (1994). Estimating terrestrial biodiversity through extrapolation. \emph{Philosophical Transactions of the Royal Society of London B - Biological Sciences} \strong{345}, 101-118
}
\examples{
data(Seedlings_Freq)
}
\keyword{datasets}