\name{Seedlings_Pre}
\alias{Seedlings_Pre}
\docType{data}
\title{
Seedlings Presence/Absence Data
}
\description{
Seedlings presence data is from Colwell and Coddington (1994) for illustration which contains 121 standardized soil samples collected from one-hundred 10 m x 10 m frids in a Coasta Rican forest.
}
\usage{data(Seedlings_Pre)}
\format{
The data is arranged in a matrix with 34 rows and 121 columns. The presence of any species in a sample is denoted by 1 and its absence is denoted by 0. \cr
A total of 34 species of seedlings germinated from these 121 samples.
}
\details{
The type of Seedlings presence/absence data is called \strong{Presence/Absence Data}. \cr
In some ecological studies, incidence (presence/absence) data are collected over repeated samples in time and/or sapce. Quadrat sampling procides an example of quadrats is randomly selected for observation.
}
\source{
Colwell, R. A., and Coddington, J. A. (1994). Estimating terrestrial biodiversity through extrapolation. \emph{Philosophical Transactions of the Royal Society of London B - Biological Sciences} \strong{345}, 101-118
}
\examples{
data(Seedlings_Pre)
}
\keyword{datasets}
