\name{Spider_data}
\alias{Spider_data}
\docType{data}
\title{
Spider data
}
\description{
Sackett et al. (2011) provided species abundance data for samples of spiders from four 5 experimental forest canopy-manipulation treatments at the Harvard Forest. \cr\cr
Data from one treatments is used here for illustration: the 9 Hemlock Girdled treatment, in which bark and cambium of hemlock trees were cut and the trees
10 left in place to die to mimic tree mortality by adelgid infestation.
}
\usage{data(Spider_data)}
\format{
  The format is: c(0, 15, 46, 2, 0, 0, 0, 1, 6, 1, ...) 
}

\source{
Sackett, T. E., S. Record, S. Bewick, B. Baiser, N. J. Sanders, & A. M. Ellison. 2011. Response of 5 macroarthropod assemblages to the loss of hemlock (Tsuga canadensis), a foundation species.
}
\examples{
data(Spider_data)
}
\keyword{datasets}
